Move methods and classes for consistency
